+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - Add new region `wetzikon` based on the provided iCal-Files
 - Add new region `wiesendangen` based on published PDF
+- Add new region `hombrechtikon` based on published PDF
 - Add new region `winterthur` based on the provided iCal-File
 - Update data of Embrach